requirements.txt
numpy
pandas
requests
joblib
scikit-learn
matplotlib
google-cloud-storage
warcio
python-dotenv
google-cloud-aiplatform